登録処理と一覧を分ける
