+++ erp/web/src/store/modules/client.ts
@@ -3,18 +3,13 @@ import { createClient, getClient, updateClient, infoClient } from '@/api/client'
 import store from '@/store'
 export interface IClientState {
   id: string,
-  name: string,
-  list: JSON[]
-  totalItem: number
+  name: string
 }
 
 @Module({ dynamic: true, store, name: 'client' })
 class Client extends VuexModule implements IClientState {
   public id = ''
   public name = ''
-  public list: JSON[] = []
-  public totalItem=0
-
   @Mutation
   private SET_ID(id: string){
     this.id = id
@@ -25,15 +20,6 @@ class Client extends VuexModule implements IClientState {
     this.name = name
   }
 
-  @Mutation
-  private SET_LIST(list: JSON[]){
-    this.list = list
-  }
-
-  @Mutation
-  private SET_TOTAL_ITEM(totalItem: number) {
-    this.totalItem = totalItem
-  }
 
   @Action
   public async CreateClient(clientInfo: { name: string }) {
@@ -49,27 +35,12 @@ class Client extends VuexModule implements IClientState {
     store.commit('SET_NAME', name)
   }
 
-  
   @Action({ rawError: true })
   public async EditClient(clientInfo: { id: string }) {
     let { id } = clientInfo
     const { data } = await getClient({ clientsSeq: id })
     this.SET_ID(data.clientsSeq)
     this.SET_NAME(data.clientsName)
-  }
-
-  @Action({ rawError: true })
-  public async UpdateClient(clientInfo: { id: string, name: string }) {
-    let {id, name } = clientInfo
-    name = name.trim()
-    await updateClient({ clientsName: name, clientsSeq: id })
-  }
-
-  @Action({ rawError: true })
-  public async ClientList(clientInfo: any){
-    const { data } = await infoClient(clientInfo)
-    this.SET_LIST(data.clients)
-    this.SET_TOTAL_ITEM(data.totalItemsNum)
   }
 }
